+++ tests/unit/stubs.py
@@ -1,3 +1,8 @@
+"""
+Note: this module is NOT listed in any settings.MODERNRPC_METHODS_MODULES
+All methods declared here must be manually registered to be available to remote call
+"""
+
 from modernrpc.core import rpc_method
 
 
@@ -27,3 +32,15 @@ def dummy_remote_procedure_4():
 
 def not_decorated_procedure():
     pass
+
+
+@rpc_method(name="foo", entry_point="v1")
+def func_v1():
+    # Function to register with the name "foo" under "v1" entry_point
+    return "V1"
+
+
+@rpc_method(name="foo", entry_point="v2")
+def func_v2():
+    # Function to register with the name "foo" under "v2" entry_point
+    return "V2"