+++ CHANGELOG.md
@@ -9,6 +9,7 @@
   - JSON-RPC handler now correctly checks type of request "id" field, according to the
     [specification](https://www.jsonrpc.org/specification#request_object) (#72). Thanks to @moonburnt
   - Fixed docstring parser to allow empty or multi-line arg description
+  - Fixed registration of methods with same names under different entry points (#74). Thanks to @hwalinga
 
 ### Misc